+++ lib/Dwoo/Plugins/Functions/PluginExtends.php
@@ -1,16 +1,16 @@
 <?php
 /**
- * Copyright (c) 2013-2016
+ * Copyright (c) 2013-2017
  *
  * @category  Library
  * @package   Dwoo\Plugins\Functions
  * @author    Jordi Boggiano <j.boggiano@seld.be>
  * @author    David Sanchez <david38sanchez@gmail.com>
  * @copyright 2008-2013 Jordi Boggiano
- * @copyright 2013-2016 David Sanchez
- * @license   http://dwoo.org/LICENSE Modified BSD License
- * @version   1.3.0
- * @date      2016-09-19
+ * @copyright 2013-2017 David Sanchez
+ * @license   http://dwoo.org/LICENSE LGPLv3
+ * @version   1.3.6
+ * @date      2017-03-21
  * @link      http://dwoo.org/
  */
 
@@ -75,8 +75,8 @@ class PluginExtends extends Plugin implements ICompilable
             self::$r = '\s*' . self::$r;
         }
         $inheritanceTree = array(array('source' => $compiler->getTemplateSource()));
-        $curPath         = dirname($compiler->getDwoo()->getTemplate()->getResourceIdentifier()) . DIRECTORY_SEPARATOR;
-        $curTpl          = $compiler->getDwoo()->getTemplate();
+        $curPath         = dirname($compiler->getCore()->getTemplate()->getResourceIdentifier()) . DIRECTORY_SEPARATOR;
+        $curTpl          = $compiler->getCore()->getTemplate();
 
         while (!empty($file)) {
             if ($file === '""' || $file === "''" || (substr($file, 0, 1) !== '"' && substr($file, 0, 1) !== '\'')) {
@@ -94,7 +94,7 @@ class PluginExtends extends Plugin implements ICompilable
             }
 
             try {
-                $parent = $compiler->getDwoo()->templateFactory($resource, $identifier, null, null, null, $curTpl);
+                $parent = $compiler->getCore()->templateFactory($resource, $identifier, null, null, null, $curTpl);
             }
             catch (SecurityException $e) {
                 throw new CompilationException($compiler, 'Extends : Security restriction : ' . $e->getMessage());